Formatting: drop the obsolete formatting services
We now rely on the default formatting.
