+++ src/main/scala/me/fornever/haskeletor/language/HaskellFormattingService.scala
@@ -6,15 +6,24 @@
 
 package me.fornever.haskeletor.language
 
+import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.formatting.service.{AsyncDocumentFormattingService, AsyncFormattingRequest, FormattingService}
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.diagnostic.Logger
+import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiFile
-import me.fornever.haskeletor.action.OrmoluReformatAction
 import me.fornever.haskeletor.core.HaskeletorBundle
-import me.fornever.haskeletor.util.HaskellFileUtil
+import me.fornever.haskeletor.external.component.StackProjectManager
+import me.fornever.haskeletor.settings.HTool
+import me.fornever.haskeletor.util.{HaskellFileUtil, ScalaUtil}
 
+import java.io.InputStream
+import java.nio.charset.StandardCharsets
+import java.nio.file.Path
 import java.util
 import java.util.concurrent.CompletableFuture
-import java.util.concurrent.atomic.AtomicReference
+import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
+import scala.io.Source
 
 final class HaskellFormattingService extends AsyncDocumentFormattingService {
 
@@ -33,7 +42,7 @@ final class HaskellFormattingService extends AsyncDocumentFormattingService {
       override def run(): Unit = {
         val maybeFuture = for {
           ioFile <- Option(formattingRequest.getIOFile)
-          future <- OrmoluReformatAction.reformatFile(formattingRequest.getContext.getProject, ioFile.toPath)
+          future <- reformatFile(formattingRequest.getContext.getProject, ioFile.toPath)
         } yield future
 
         maybeFuture match {
@@ -62,5 +71,87 @@ final class HaskellFormattingService extends AsyncDocumentFormattingService {
       override def cancel(): Boolean = Option(runningFuture.get()).exists(_.cancel(true))
 
       override def isRunUnderProgress: Boolean = true
+    }
+
+  private def reformatFile(project: Project, file: Path): Option[CompletableFuture[String]] = {
+    StackProjectManager.isOrmoluAvailable(project).map { ormoluPath =>
+      val commandLine = new GeneralCommandLine(
+        ormoluPath,
+        file.toString
+      )
+
+      val processRef = new AtomicReference[Process]()
+      val cancelled = new AtomicBoolean(false)
+
+      val future = new CompletableFuture[String]() {
+        override def cancel(mayInterruptIfRunning: Boolean): Boolean = {
+          val wasCancelled = super.cancel(mayInterruptIfRunning)
+          if (wasCancelled) {
+            cancelled.set(true)
+            Option(processRef.get()).foreach { process =>
+              process.destroy()
+              if (process.isAlive) {
+                process.destroyForcibly()
+              }
+            }
+          }
+          wasCancelled
+        }
+      }
+
+      ApplicationManager.getApplication.executeOnPooledThread(ScalaUtil.runnable {
+        try {
+          val process = commandLine.createProcess()
+          processRef.set(process)
+
+          if (future.isCancelled || cancelled.get()) {
+            process.destroy()
+            if (process.isAlive) {
+              process.destroyForcibly()
+            }
+          } else {
+            val stdoutFuture = ApplicationManager.getApplication.executeOnPooledThread(ScalaUtil.callable[String] {
+              readStream(process.getInputStream)
+            })
+            val stderrFuture = ApplicationManager.getApplication.executeOnPooledThread(ScalaUtil.callable[String] {
+              readStream(process.getErrorStream)
+            })
+
+            val exitCode = process.waitFor()
+            val stdout = stdoutFuture.get()
+            val stderr = stderrFuture.get()
+
+            if (exitCode == 0) {
+              future.complete(stdout)
+            } else {
+              logger.error(
+                s"Ormolu reformat process failed for `$file` with exit code $exitCode.\nstdout:\n$stdout\nstderr:\n$stderr"
+              )
+              val truncatedStderr = stderr.take(1024)
+              future.completeExceptionally(
+                new RuntimeException(
+                  s"Error while reformatting by `${HTool.Ormolu.name}`. Exit code: $exitCode. Error: $truncatedStderr"
+                )
+              )
+            }
+          }
+        } catch {
+          case e: Throwable =>
+            if (!future.isCancelled && !cancelled.get()) {
+              future.completeExceptionally(e)
+            }
+        }
+      })
+
+      future
+    }
+  }
+
+  private val logger = Logger.getInstance(getClass)
+
+  private def readStream(stream: InputStream): String = {
+    val source = Source.fromInputStream(stream, StandardCharsets.UTF_8.name())
+    try source.mkString
+    finally source.close()
   }
 }
