--- vcs/src/main/kotlin/me/fornever/haskeletor/vcs/HaskellReformatCheckinHandlerFactory.kt
@@ -1,25 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2014-2022 Rik van der Kleij
- * SPDX-FileCopyrightText: 2014-2022 intellij-haskell contributors <https://github.com/rikvdkleij/intellij-haskell>
- * SPDX-FileCopyrightText: 2026 haskeletor contributors <https://github.com/ForNeVeR/haskeletor>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package me.fornever.haskeletor.vcs
-
-import com.intellij.openapi.vcs.CheckinProjectPanel
-import com.intellij.openapi.vcs.changes.CommitContext
-import com.intellij.openapi.vcs.checkin.CheckinHandler
-import com.intellij.openapi.vcs.checkin.CheckinHandlerFactory
-import me.fornever.haskeletor.projectmodel.HaskellProjectManager
-
-class HaskellReformatCheckinHandlerFactory : CheckinHandlerFactory() {
-    override fun createHandler(panel: CheckinProjectPanel, commitContext: CommitContext): CheckinHandler {
-        return if (HaskellProjectManager.getInstance(panel.project).isHaskellProject.valueOrNull == true) {
-            HaskellReformatBeforeCheckinHandler(panel.project, panel)
-        } else {
-            CheckinHandler.DUMMY
-        }
-    }
-}
